Added more instructions and the putc system call. Changed the Makefile.
As I'm emulating the user mode at present, I added a putc() system call
so that I could call $display in Verilog to get some output to happen.

I also rewrote some of the Makefile to use the new start_swivm program
to read and convert the output from the Swieros C compiler.